[TASK] Relocate viewhelper and add alias
